Add basic now ls support with projects.
